Move getting started service to workbench/contrib Closes #118603
