+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/colorTheme.png
