+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/commandPalette.png
