+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/dark/settings.png
