+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/extensions.png
