+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/github.png
