+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/languageExtensions.png
