+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/openFolder.png
