+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/openVSC.png
