+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/pullRequests.png
