+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/remoteTerminal.png
