+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/runProject.png
