+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/settingsSync.png
