+++ src/vs/workbench/contrib/welcome/gettingStarted/common/media/terminal.png
